refactor(core): modularize `decorators.ts` into individual files and improve package structure
Split the monolithic `decorators.ts` into modular files for better structure and maintainability. Adjusted `tsconfig` includes and added a central `index.ts` for exports. Updated metadata keys and package paths to ensure seamless integration. Incremented package versions to `4.0.1`.
