+++ .aiignore
@@ -4,3 +4,4 @@ node_modules
 dist
 node_modules
 bun.lock
+references
