+++ .gitignore
@@ -37,3 +37,4 @@ report.[0-9]_.[0-9]_.[0-9]_.[0-9]_.json
 lcov
 /.env.dev
 /examples/packages/cf-worker/worker-configuration.d.ts
+/references/
