+++ packages/di-framework-cli/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@di-framework/cli",
-  "version": "4.0.0",
+  "version": "4.0.1",
   "private": false,
   "type": "module",
   "main": "main.ts",
