+++ packages/di-framework-core/decorators/keys.ts
@@ -1,0 +1,9 @@
+/**
+ * Metadata keys shared between the decorators in this folder.
+ *
+ * These mirror the private constants in `../container.ts`; the string values
+ * must stay in sync so the container can read what the decorators write.
+ */
+
+export const INJECTABLE_METADATA_KEY = 'di:injectable';
+export const INJECT_METADATA_KEY = 'di:inject';
