+++ packages/di-framework-repo/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@di-framework/repo",
-  "version": "4.0.0",
+  "version": "4.0.1",
   "description": "A coherent abstraction of repositories and storage adapters for di-framework.",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
